video_rag/requirements.txt
streamlit
google-generativeai
opencv-python-headless
ffmpeg-python
yt-dlp
pydub
pandas
selenium
beautifulsoup4
webdriver-manager
openai-whisper
youtube-transcript-api
langchain
langchain-community
langchain-google-genai
chromadb
lxml
